requirements.txt
streamlit
replicate
pillow
requests
streamlit-compare-images